Factors a lot of code out of correctBiasStored/correctBias.py
This is in preparation for also factoring it out of correctBias/correctBias.py,
eliminating a lot of code duplication between the two modules.
