+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = CRADLE
-version = 0.19.0
+version = 0.20.0
 description = Correct Read Counts and Analysis of Differently Expressed Regions
 long_description = file: README.md
 long_description_content_type = text/markdown
@@ -19,6 +19,7 @@ packages =
     CRADLE.CorrectBias
     CRADLE.CorrectBiasStored
     CRADLE.CallPeak
+    CRADLE.correctbiasutils
 package_dir =
     CRADLE=CRADLE
 zip_safe = False
