+++ setup.py
@@ -17,6 +17,10 @@ ext_modules = [
 	Extension(
 		'CRADLE.CallPeak.calculateRC',
 		['CRADLE/CallPeak/calculateRC.pyx'],
+		extra_compile_args=["-fno-strict-aliasing"]),
+	Extension(
+		'CRADLE.correctbiasutils.cython',
+		['CRADLE/correctbiasutils/cython.pyx'],
 		extra_compile_args=["-fno-strict-aliasing"])
 ]
